Fixed Dumb Code
Fixed some stupid code that called a planet event on every planet in a country, which would lag the game out. Adding to that, the "planet event" being called is really a country event that already has every_owned_planet. AlphaAsh, folk, Crusader Vanguard, and pretty much every other modder would like to add that "whoever wrote this is a dumb fuck".